refactor: organize classes and types
