--- src/api/SearchResults.ts
@@ -1,4 +1,0 @@
-export interface SearchResults {
-    items?: object[];
-    totalResults?: number;
-}
