+++ src/api/index.ts
@@ -1,0 +1,3 @@
+export * from './search'
+export * from './search.types'
+export * from './ResultsCache'
