+++ src/api/search.types.ts
@@ -1,0 +1,8 @@
+export interface SearchQuery {
+    query: string;
+}
+
+export interface SearchResults {
+    items?: Array<object>;
+    totalResults?: number;
+}
